# Add (de)serialization support for Base64 string as Byte array