--- a/src/System.Text.Json/src/System/Text/Json/Serialization/JsonPropertyInfoCommon.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/JsonPropertyInfoCommon.cs
@@ -50,7 +50,6 @@ namespace System.Text.Json.Serialization
                 IsPropertyPolicy = true;
                 HasGetter = true;
                 HasSetter = true;
-                ValueConverter = DefaultConverters<TRuntimeProperty>.s_converter;
             }
 
             GetPolicies();