--- a/src/System.Text.Json/tests/Serialization/TestClasses.SimpleTestClass.cs
+++ b/src/System.Text.Json/tests/Serialization/TestClasses.SimpleTestClass.cs
@@ -111,7 +111,7 @@ namespace System.Text.Json.Serialization.Tests
                 @"""MyUInt16Array"" : [4]," +
                 @"""MyUInt32Array"" : [5]," +
                 @"""MyUInt64Array"" : [6]," +
-                @"""MyByteArray"" : [7]," +
+                @"""MyByteArray"" : ""Bw==""," + // Base64 encoded value of 7
                 @"""MySByteArray"" : [8]," +
                 @"""MyCharArray"" : [""a""]," +
                 @"""MyStringArray"" : [""Hello""]," +