Uptake the recent Base64 reader\writer work to the serializer.

Addresses serialization work for #36974 and follow up for [reader\writer\doc PR](https://github.com/dotnet/corefx/pull/38088).